test3/test3_gxr/requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.21
Pillow>=9.0
matplotlib>=3.5
